% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{fetchClinvar}
\alias{fetchClinvar}
\title{Fetch missense variants in ClinVar for given gene}
\usage{
fetchClinvar(gene)
}
\arguments{
\item{gene}{gene name as in ClinVar (e.g. 'CALM1')}
}
\value{
a \code{data.frame} with the following columns:
\itemize{
  \item \code{hgvsc} The HGVS variant descriptor at the coding sequence (DNA) level
  \item \code{hgvsp} The HGVS variant descriptor at the amino acid sequence (Protein) level
  \item \code{clinsig} The ClinVar clinical significance string (e.g "Likely pathogenic")
}
}
\description{
Makes a query to the NCBI webservice to fetch the ClinVar entries for the given gene
and filters them down to missense variants
}
