% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{htmlDecode}
\alias{htmlDecode}
\title{Decode HTML strings}
\usage{
htmlDecode(str)
}
\arguments{
\item{str}{input string}
}
\value{
the decoded string
}
\description{
Decode HTML strings by converting special character escape sequences (e.g. \code{&gt;})
to their corresponding UTF-8 characters.
}
