% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{map2bf.file}
\alias{map2bf.file}
\title{Calibrate a variant map from a file to Bayes Factors}
\usage{
map2bf.file(csvfile, ensembls, symbols, drawPlot = TRUE)
}
\arguments{
\item{csvfile}{the scoreset file in csv format}

\item{drawPlot}{boolean value indicating whether a plot of variant distributions
should be drawn. Defaults to \code{TRUE}.}

\item{ensembl}{the ensembl ID for the target gene}

\item{symbol}{the gene symbol of the target gene}
}
\value{
a \code{data.frame} containing the original \code{scores} input
  with an additional column \code{llr}, representing the log likelihood ratio
  (i.e. log Bayes Factor)
}
\description{
Calibrate a variant map from a file to Bayes Factors
}
