% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{map2bf}
\alias{map2bf}
\title{Calibrate a variant map to Bayes Factors}
\usage{
map2bf(scores, ensembls, symbols, drawPlot = TRUE)
}
\arguments{
\item{scores}{a dataframe with columns: hgvs_nt, hgvs_pro, score}

\item{ensembls}{a vector of ensembl IDs for the target gene}

\item{symbols}{a vector of gene symbols of the target gene}

\item{drawPlot}{boolean value indicating whether a plot of variant distributions
should be drawn. Defaults to \code{TRUE}.}
}
\value{
a \code{data.frame} containing the original \code{scores} input
  with an additional column \code{llr}, representing the log likelihood ratio
  (i.e. log Bayes Factor)
}
\description{
Calibrate a variant map to Bayes Factors
}
