% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{new.family}
\alias{new.family}
\title{New Family Object}
\usage{
new.family(maf, etiologic.fraction, penetrance, self = c("heterozygous",
  "homozygous"), modus = c("dominant", "recessive"))
}
\arguments{
\item{maf}{Minor Allele Frequency of allele in question}

\item{etiologic.fraction}{The fraction of symptomatic patients that carry the genotype}

\item{penetrance}{The fraction of genotype carriers that are symptomatic}

\item{self}{The genotype of the patient him/herself. "heterozyogus" or "homozygous"}

\item{modus}{The mode of inheritance. "dominant" or "recessive"}
}
\value{
a new Family object
}
\description{
Creates a new family object, allowing for calculation of a Bayes factor. The Bayes Factor
Represents the likelyhood ratio of a phenotype being caused by an allele in question.
}
\details{
A family object offers the following functions:
\itemize{
  \item \code{add.parent(pheno)} Adds a parent to the family, with boolean parameter \code{pheno}
     indicating whether the parent has the phenotype/disease.
  \item \code{add.sibling(pheno)} Adds a sibling to the family, with boolean parameter \code{pheno}
     indicating whether the sibling has the phenotype/disease.
  \item \code{add.child(pheno)} Adds a child to the family, with boolean parameter \code{pheno}
     indicating whether the child has the phenotype/disease.
  \item \code{log.bayes.factor()} Returns the overall log likelihood ratio (i.e. log Bayes-Factor) 
     for this family.
}
}
\examples{
family <- new.family(
  maf=0.05,
  etiologic.fraction=0.8,
  penetrance=0.8,
  self="heterozygous",
  modus="dominant"
)
family$add.parent(pheno=TRUE)
family$add.parent(pheno=FALSE)
family$add.sibling(pheno=TRUE)
family$summary()
}
