% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotInterval}
\alias{plotInterval}
\title{Gradient plot for LLR and probability confidence intervals}
\usage{
plotInterval(lower, mid, upper, type = c("llr", "probability"),
  resolution = 50)
}
\arguments{
\item{lower}{the lower value of the CI}

\item{mid}{the maximum likelihood value of the CI}

\item{upper}{the upper value of the CI}

\item{type}{the type of value used (either llr or probability). 
This determines the range and gradient underlying the plot}
}
\description{
Draws a gradient plot for LLR and probability confidence intervals
}
