% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{fetchGnomad}
\alias{fetchGnomad}
\title{Fetch missense variants in GnomAD for given gene}
\usage{
fetchGnomad(ensemblID, overrideCache = FALSE, logger = NULL)
}
\arguments{
\item{ensemblID}{The Ensembl gene identifier (e.g. \code{ENSG00000198668})}

\item{overrideCache}{logical determining whether local cache should be overridden}

\item{logger}{a yogilogger object to write log messages to. Defaults to NULL}
}
\value{
a \code{data.frame} with the following columns:
\itemize{
  \item \code{hgvsc} The HGVS variant descriptor at the coding sequence (DNA) level
  \item \code{hgvsp} The HGVS variant descriptor at the amino acid sequence (Protein) level
  \item \code{maf} The minor allele frequency.
}
}
\description{
Makes a query to the ExAC webservice to fetch the GnomAD entries for the given gene
and filters them down to missense variants.
}
