% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{new.persistence.connection}
\alias{new.persistence.connection}
\title{Create a new persistence connection object}
\usage{
new.persistence.connection(dbfile)
}
\arguments{
\item{dbfile}{the SQLite database file}
}
\value{
a new persistence.connection object
}
\description{
Constructor for creating a new object of type persistence.connection.
}
\details{
Provides the following methods:
\begin{description}
\item{isKnown(urn)} checks whether a URN is known to the database
\item{getStatus(urn)} retrieves the status of the scoreset with the given URN
\item{setStatus(urn)} sets the status of the scoreset with the given URN to the given status. 
    Permissible values are "new", "configured", "calibrated".
\item{getParameters(urn)} retrieves the calibration parameters for the scoreset with the given URN
\item{newScoreset(urn,symbols,ensembl)} adds a new scoreset with the given URN to the database and initializes
    its gene symbols and ensembl gene names to the given values.
\item{close()} closes the database connection
\end{description}
}
\examples{
dbfile <- "maveclin.db"
persist <- new.persistence.connection(dbfile)
persist$isKnown("urn:mavedb:00000001-c-2")
persist$getStatus("urn:mavedb:00000001-c-2")
persist$setStatus("urn:mavedb:00000001-c-2","new")
persist$close()

}
